refactor protocolMetrics update triggers
