+++ src/OtterQiLocker.ts
@@ -5,7 +5,6 @@ import {
 import { updateTreasuryRevenueHarvest } from './utils/TreasuryRevenue'
 import { Harvest } from '../generated/schema'
 import { loadOrCreateTransaction } from './utils/Transactions'
-import { updateProtocolMetrics } from './utils/ProtocolMetrics'
 
 export function handleHarvest(event: HarvestEvent): void {
   let transaction = loadOrCreateTransaction(event.transaction, event.block)
@@ -14,7 +13,6 @@ export function handleHarvest(event: HarvestEvent): void {
   entity.timestamp = transaction.timestamp
   entity.amount = event.params.amount
   updateTreasuryRevenueHarvest(event.block.number, entity)
-  updateProtocolMetrics(transaction)
   entity.save()
 }
 
@@ -25,6 +23,5 @@ export function handleHarvestStaking(event: HarvestStakingEvent): void {
   entity.timestamp = transaction.timestamp
   entity.amount = event.params.amount
   updateTreasuryRevenueHarvest(event.block.number, entity)
-  updateProtocolMetrics(transaction)
   entity.save()
 }
